% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{addGenomes}
\alias{addGenomes}
\title{Add new organisms to an existing pangenome}
\usage{
addGenomes(object, newSet, ...)
}
\arguments{
\item{object}{A pgVirtual subclass to merge the new genomes into}

\item{newSet}{An object of the same class as object containing the new
organisms to add. Grouping of the genes contained in this object can already
exist, if not it will be done automatically.}

\item{...}{parameters passed on to community detection algorithm.}
}
\value{
An object of the same class as object containing the new organisms
from newSet and possible new gene groups from genes with no orthologues in
the original pangenome.
}
\description{
This method allows new genomes to be added to an already processed pangenome,
preserving existing grouping and adding new genes to their relevant groups.
This makes it possible to gradually grow the pangenome as new sequences
becomes available without redoing the grouping at each time, loosing the gene
group metadata.
}

