% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modifying.R
\docType{methods}
\name{addGroupInfo}
\alias{addGroupInfo}
\alias{addGroupInfo,pgVirtual-method}
\title{Safely add group info}
\usage{
addGroupInfo(object, ...)

\S4method{addGroupInfo}{pgVirtual}(object, info, key)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on.}

\item{info}{A data.frame with information to add}

\item{key}{Either an integer vector with the index of each gene group the 
rows in info corresponds to, or the name of the column in info that holds the
indexes.}
}
\value{
An object of the same class as object with the new gene group 
information.
}
\description{
This method allows for adding of group metadata by specifying the name of the
metadata and the gene groups it should be added to. It protects the user from
overwriting information that is derived from the data, and ensures the proper
formatting. Should be prefered to \code{\link{groupInfo<-}} for all but the
simplest cases.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Add gene group info safely for all pgVirtual 
subclasses
}}

\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Create some info
info <- data.frame(nickname=c('Tessie', 'Johnny'), index=c(4, 500))

# Add it to the object
testPG <- addGroupInfo(testPG, info=info, key='index')

}
\seealso{
Other Metadata: \code{\link{addOrgInfo}},
  \code{\link{groupInfo}}, \code{\link{orgInfo}}
}
\concept{Metadata}
