% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modifying.R
\docType{methods}
\name{addOrgInfo}
\alias{addOrgInfo}
\alias{addOrgInfo,pgVirtual-method}
\title{Safely add organisms info}
\usage{
addOrgInfo(object, ...)

\S4method{addOrgInfo}{pgVirtual}(object, info, key)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on.}

\item{info}{A data.frame with information to add}

\item{key}{Either an integer vector with the index of each organism the 
rows in info corresponds to, or the name of the column in info that holds the
indexes.}
}
\value{
An object of the same class as object with the added organism 
information.
}
\description{
This method allows for adding of organism metadata by specifying the name of
the metadata and the organisms it should be added to. It protects the user 
from overwriting information that is derived from the data and ensures proper
formatting. Should be prefered to \code{\link{orgInfo<-}} for all but the 
simplest cases.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Add organism info safely for all pgVirtual subclasses
}}

\examples{
testPG <- .loadPgExample()

# Create some information
info <- data.frame(location=c('Copenhagen', 'Paris', 'London'), 
    name=c('AE017243', 'AP012303', 'AE017244')
)

# Add the information
testPG <- addOrgInfo(testPG, info=info, key='name')

}
\seealso{
Other Metadata: \code{\link{addGroupInfo}},
  \code{\link{groupInfo}}, \code{\link{orgInfo}}
}
\concept{Metadata}
