% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grouping.R
\docType{methods}
\name{cdhitGrouping}
\alias{cdhitGrouping}
\alias{cdhitGrouping,pgVirtual-method}
\title{Gene grouping by preclustering with CD-HIT}
\usage{
cdhitGrouping(object, ...)

\S4method{cdhitGrouping}{pgVirtual}(object, kmerSize, lowerLimit, maxLengthDif,
  geneChunkSize, cdhitOpts)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on.}

\item{kmerSize}{The size of the kmer's used for the comparison. If two values
are given the first will be used for the CD-HIT algorithm and the second will
be used for the cosine similarity calculations.}

\item{lowerLimit}{A numeric giving the lower bounds of similarity below which
it will be set to zero.}

\item{maxLengthDif}{The maximum deviation in sequence length to allow during
preclustering with CD-HIT. Below 1 it describes a percentage. Above 1 it 
describes a fixed length.}

\item{geneChunkSize}{The maximum number of genes to pass to the CD-HIT
algorithm. If object contains more genes than this, CD-HIT will be run in 
chunks and combined with a second CD-HIT pass before the final cosine 
similarity grouping.}

\item{cdhitOpts}{Additional arguments passed on to CD-HIT. It should be a 
named list with names corresponding to the arguments expected in the CD-HIT
algorithm (without the dash). i, n and s/S will be overwritten based on the
other parameters given to this function and all values in cdhitOpts will be
converted to character using as.character}
}
\value{
An object of the same class as 'object'.
}
\description{
This grouping algorithm partly mimicks the approach used by Roary, but
instead of using BLAST in the second pass it uses cosine similarity of kmer
feature vectors, thus providing an even greater speedup. The algorithm uses
the CD-HIT algorithm to precluster highly similar sequences and then groups
these clusters by extracting a representative and clustering these using the
standard FindMyFriends kmer cosine similarity.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Grouping using cdhit for all pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample()

testPG <- cdhitGrouping(testPG)

}
\references{
Page, A. J., Cummins, C. A., Hunt, M., Wong, V. K., Reuter, S., Holden, M. T. 
G., et al. (2015). Roary: rapid large-scale prokaryote pan genome analysis. 
\emph{Bioinformatics}, btv421.

Fu, L., Niu, B., Zhu, Z., Wu, S., Li, W. (2012). CD-HIT: 
accelerated for clustering the next generation sequencing data. 
\emph{Bioinformatics}, \bold{28} (23), 3150--3152.

Li, W. and Godzik, A. (2006) Cd-hit: a fast program for clustering and 
comparing large sets of protein or nucleotide sequences. 
\emph{Bioinformatics}, \bold{22}, 1658--9.
}
\seealso{
Other grouping algorithms: \code{\link{gpcGrouping}},
  \code{\link{graphGrouping}}, \code{\link{manualGrouping}}
}

