% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modifying.R
\docType{methods}
\name{collapseParalogues}
\alias{collapseParalogues}
\alias{collapseParalogues,pgVirtual-method}
\title{Merge paralogue gene groups into new gene groups}
\usage{
collapseParalogues(object, ...)

\S4method{collapseParalogues}{pgVirtual}(object, combineInfo = "merge",
  ...)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on to metadata collapse function. For 
combineInfo='merge' sep specifies the separator - sep='none' collapses 
information into list elements instead of strings. For combineInfo='largest'
no addition arguments are given.}

\item{combineInfo}{The approach used to combine metadata from the collapsed
groups. Either 'merge' for merging, 'largest' for picking information from
the largest group, or a function that takes a data.frame of multiple rows and
converts it to a data.frame with one row and the same columns.}
}
\value{
An object of the same class as object with the new grouping.
}
\description{
This method allows for merging of paralogue gene groups defined using 
\code{\link{kmerLink}} into new, bigger, gene groups.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Merge paralogue gene groups for all pgVirtual
subclasses
}}

\examples{
testPG <- .loadPgExample(withGroups=TRUE, withParalogues=TRUE)

# Number of gene groups before collapse
nGeneGroups(testPG)

# Number of gene groups after collapse
testPG <- collapseParalogues(testPG, combineInfo='largest')
nGeneGroups(testPG)

}
