% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.fillDefaults}
\alias{.fillDefaults}
\title{Assign object defaults to missing values}
\usage{
.fillDefaults(def)
}
\arguments{
\item{def}{A named list of default values}
}
\value{
This function is called for its side effects
}
\description{
This function takes care of investigating the enclosing functions arguments
and identifying the missing ones. If they are missing and a default is given
this value is assigned to the enclosing functions environment
}
\examples{
# Should only be called within methods/functions

# This will obviously fail
\dontrun{
  t <- function(x) {
    x+1
  }
  t()
}

# Using .fillDefaults
t <- function(x, defs) {
  .fillDefaults(defs)
  x+1
}

# With defaults
t(defs=list(x=5))

# Direct setting takes precedence
t(x=2, defs=list(x=5))

# Still fails if defs doesn't contain the needed parameter
\dontrun{
  t(defs=list(y='no no'))
}

# Usually defs are derived from the object in a method:
\dontrun{
  setMethod('fillDefExample', 'pgFull',
    function(object, x, y) {
      .fillDefaults(defaults(object))
      x+y
    }
  )
}

}
\seealso{
Set and get pangenome defaults with \code{\link{defaults}}
}
