% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMemLoc.R
\docType{methods}
\name{geneLocation}
\alias{geneLocation}
\alias{geneLocation,pgInMemLoc-method}
\title{Get gene location for all genes}
\usage{
geneLocation(object)

\S4method{geneLocation}{pgInMemLoc}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A data.frame as described above
}
\description{
This method returns the gene location of all genes as a data.frame with each
row corresponding to a gene in the pangenome. The data.frame will have the 
columns 'start', 'end', 'contig' and 'strand' (order of columns not ensured)
with start and end giving the start and end position of the gene on the 
contig/chromosome given in the contig column. Strand gives the direction of
translation, 1 is from start to end and -1 is from end to start (thus start 
should always be lower than end no matter the direction of translation)
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMemLoc}: Get gene location for pgInMemLoc subclasses
}}
\note{
Required for subclasses of pgVirtualLoc in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample(geneLoc=TRUE)
head(geneLocation(testPG))

}

