% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgLM.R, R/pgFull.R, R/pgSlim.R
\docType{methods}
\name{geneNames}
\alias{geneNames}
\alias{geneNames,pgFull-method}
\alias{geneNames,pgLM-method}
\alias{geneNames<-}
\alias{geneNames<-,pgFull-method}
\alias{geneNames<-,pgLM-method}
\title{Get and set the names of the genes in the pangenome}
\usage{
geneNames(object)

geneNames(object) <- value

\S4method{geneNames}{pgLM}(object)

\S4method{geneNames}{pgLM}(object) <- value

\S4method{geneNames}{pgFull}(object)

\S4method{geneNames}{pgFull}(object) <- value

\S4method{geneNames}{pgLM}(object)

\S4method{geneNames}{pgLM}(object) <- value
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{value}{A character vector with new names}
}
\value{
In case of the getter, a character vector containing the names of 
each gene.
}
\description{
These methods lets you query and change the naming of genes in your 
pangenome. Take note that even though sequences are not in memory for pgLM
objects, the names are. This means that changes to the description header in
the underlying fasta files have no effect on the naming in your pangenome
}
\section{Methods (by class)}{
\itemize{
\item \code{pgLM}: Get genenames for pgLM and subclasses

\item \code{pgLM}: Set genenames for pgLM and subclasses

\item \code{pgFull}: Get genenames for pgFull and subclasses

\item \code{pgFull}: Set genenames for pgFull and subclasses

\item \code{pgLM}: Throws error for pgSlim

\item \code{pgLM}: Throws error for pgSlim
}}
\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample()
head(geneNames(testPG))

geneNames(testPG)[10] <- 'Gene number 10'

}

