% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgLM.R, R/pgFull.R
\docType{methods}
\name{geneWidth}
\alias{geneWidth}
\alias{geneWidth,pgFull-method}
\alias{geneWidth,pgLM-method}
\title{Get the sequence length of each gene}
\usage{
geneWidth(object)

\S4method{geneWidth}{pgLM}(object)

\S4method{geneWidth}{pgFull}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
An integer vector with the length of each sequence
}
\description{
This method extracts the width (i.e. number of residues) of each gene in
the pangenome.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgLM}: Get gene width for pgLM and subclasses

\item \code{pgFull}: Get gene widths for pgFull and subclasses
}}
\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample()
head(geneWidth(testPG))

}

