% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/investigating.R
\docType{methods}
\name{getNeighborhood}
\alias{getNeighborhood}
\alias{getNeighborhood,pgVirtualLoc-method}
\title{Extract a graph representation of a gene group neighborhood}
\usage{
getNeighborhood(object, ...)

\S4method{getNeighborhood}{pgVirtualLoc}(object, group, vicinity = 4)
}
\arguments{
\item{object}{A pgVirtualLoc subclass}

\item{...}{Parameters passed on.}

\item{group}{Either the name or the index of the group whose neighborhood is
of interest}

\item{vicinity}{An integer giving the number of gene groups in both 
directions to collect}
}
\value{
An igraph object with gene groups as vertices and positional 
connections as edges. The edges is weighted according to the number of genes 
sharing the connection. All vertices have a centerGroup attribute, which is 
FALSE for all but the center group.
}
\description{
This method creates a graph representation of the imidiate neighborhood of
a gene group. It is different from creating a subgraph of the panchromosome
in that only vertices and edges directly reachable from the gene group is 
included. The vertices will be annotated with a centerGroup property 
indicating whether or not the node is the queried gene group.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtualLoc}: Gene group neighborhoods for all pgVirtualLoc
subclasses
}}
\examples{
testPG <- .loadPgExample(geneLoc=TRUE, withNeighborhoodSplit=TRUE)

# Look at the surroundings of group 10
neighborhood <- getNeighborhood(testPG, group=10)

}
\seealso{
\code{\link{plotNeighborhood}} for nice plotting of the neighborhood
}

