% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{getRep}
\alias{getRep}
\alias{getRep,pgVirtual,character-method}
\title{Get a representative sequence for each gene group}
\usage{
getRep(object, method)

\S4method{getRep}{pgVirtual,character}(object, method)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{method}{The method to use to get a representative. Either 'random',
'shortest' or 'longest'.}
}
\value{
An XStringSet
}
\description{
This method returns a representative sequence for each of the gene groups
defined in the pangenome. Currently the methods defined for selecting
sequences are 'random', 'shortest', and 'longest. In case of tie for the two
latter the first occurence gets returned. Consensus sequence might be added
at a latter stage.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = pgVirtual,method = character}: Get a representative sequence for each gene group for
pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Get a random sequence from each group
getRep(testPG, 'random')
}

