% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parsers.R
\name{gff2StringSet}
\alias{gff2StringSet}
\title{Read gff file into DNAStringSet}
\usage{
gff2StringSet(file, type = "CDS")
}
\arguments{
\item{file}{The path to a GFF file.}

\item{type}{Character vector or NULL. The feature types to include in the
output. NULL means include all.}
}
\value{
A DNAStringSet with the sequences of the features annotated in the
GFF file.
}
\description{
This function parses a gff file into a DNAStringSet provided that sequence
information is included in the bottom of the file (not a requirement for gff
files). The features can be filtered upfront by type (such as CDS and exon).
GFF annotation information is stored alongside the DNAStringSet in the
elementMetadata slot as a DataFrame (unfortunately it is not possible to
assign GRanges object to elementMetadata).
}

