% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/grouping.R
\docType{methods}
\name{gpcGrouping}
\alias{gpcGrouping}
\alias{gpcGrouping,pgVirtual-method}
\title{Guided Pairwise Comparison  grouping of genes}
\usage{
gpcGrouping(object, ...)

\S4method{gpcGrouping}{pgVirtual}(object, lowMem, kmerSize, tree, lowerLimit,
  pParam, cacheDB)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on to the community detection algorithm.}

\item{kmerSize}{The size of the kmer's used for the comparison. If two values
are given and the 'tree' argument is missing, the second value is used for
tree generation. If only one value is given it is recycled.}

\item{tree}{An optional tree of class dendrogram (or that can be coerced to
one) to guide the recursive algorithm. If none is supplied it will be
generated by clustering the organisms by their total kmer numbers (summing up
for each of their genes).}

\item{lowerLimit}{A numeric giving the lower bounds of similarity below which
it will be set to zero.}

\item{pParam}{An optional BiocParallelParam object that defines the workers
used for parallelisation.}

\item{cacheDB}{A \code{\linkS4class{filehash}} object or a path to a
directory where cached results should be stored. If omitted caching will not
be done. Highly recommended for long running instances.}
}
\value{
An object of the same class as 'object'.
}
\description{
This algorithm recursively builds up a pangenome by merging subpangenomes.
The recursion follows either a supplied hierarchical clustering or one
created using kmer comparison for the full organism. At each step a
representative for each gene group is selected randomly as a representative
and gets compared to all other representatives. Gene groups are then merged
based on the pangenome created for the representatives. Due to the sampling
of representatives at each step there is a certain randomness to the
algorithm. Results should be fairly stable though, as gene groups are
compared multiple times.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: gpc grouping for all pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample()

testPG <- gpcGrouping(testPG)
}
\seealso{
Other grouping algorithms: \code{\link{graphGrouping}},
  \code{\link{graphGrouping,pgVirtual-method}};
  \code{\link{manualGrouping}},
  \code{\link{manualGrouping,pgVirtual,integer-method}},
  \code{\link{manualGrouping,pgVirtual,list-method}}
}

