% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{groupInfo}
\alias{groupInfo}
\alias{groupInfo,pgInMem-method}
\alias{groupInfo<-}
\alias{groupInfo<-,pgInMem-method}
\title{Get and set information about gene group}
\usage{
groupInfo(object)

groupInfo(object) <- value

\S4method{groupInfo}{pgInMem}(object)

\S4method{groupInfo}{pgInMem}(object) <- value
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{value}{A data.frame with a row for each group}
}
\value{
In case of the getter a data.frame with organism information.
}
\description{
These methods lets you access the information stored about each gene group 
and add to it or modify it. Upfront the following columns are present: 
'description', 'group', 'paralogue', 'GO', 'EC', 'nOrg' and 'nGenes'. All 
except 'group', 'nOrg' and 'nGenes' are filled with NA as default. The latter
are prefilled with information derived from the grouping itself and should 
not be modified manually. 'description' is meant to contain a human readable
description of the functionality of the gene group, 'GO' should contain GO 
terms (stored in a list of character vectors) and EC should contain enzyme
numbers (again stored as a list of character vectors). There is no check for 
the validity of the content so it is up to the user to ensure that the terms
added are valid. Additional columns can be added at will.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Get gene group metadata for pgInMem subclasses

\item \code{pgInMem}: Set gene group metadata for pgInMem subclasses
}}
\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)
head(groupInfo(testPG))

groupInfo(testPG)$description[1] <- 'transposase'

}
\seealso{
Other Metadata: \code{\link{addGroupInfo}},
  \code{\link{addOrgInfo}}, \code{\link{orgInfo}}
}

