% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{groupNames}
\alias{groupNames}
\alias{groupNames<-}
\alias{groupNames,pgInMem-method}
\alias{groupNames<-,pgInMem-method}
\title{Get and set the names of gene groups in the pangenome}
\usage{
groupNames(object)

groupNames(object) <- value

\S4method{groupNames}{pgInMem}(object)

\S4method{groupNames}{pgInMem}(object) <- value
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{value}{A vector with new names - will be coerced to characters}
}
\value{
In case of the getter a character vector with names
}
\description{
These methods lets you manipulate the naming of gene groups in the 
pangenome. By default organisms are numbered consecutively but this can be
changed at will. New gene groups will be numbered though despite what naming
scheme has been introduced before.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Get gene group names for pgInMem subclasses

\item \code{pgInMem}: Set gene group names for pgInMem subclasses
}}

\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)
head(groupNames(testPG))

groupNames(testPG)[20] <- 'Gene group 20'

}
