% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{hasGeneGroups}
\alias{hasGeneGroups}
\alias{hasGeneGroups,pgVirtual-method}
\title{Check whether gene groups are defined}
\usage{
hasGeneGroups(object)

\S4method{hasGeneGroups}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A boolean indicating whether gene groups have been defined (TRUE) or
not (FALSE)
}
\description{
This method checks whether any grouping of genes has been done on the
object and returns TRUE if that is the case.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Gene group check for pgVirtual subclasses
}}
\examples{
# Empty pangenome
testPG <- .loadPgExample()
hasGeneGroups(testPG)

# With gene groups
testPG <- .loadPgExample(withGroups=TRUE)
hasGeneGroups(testPG)
}

