% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{hasGeneInfo}
\alias{hasGeneInfo}
\alias{hasGeneInfo,pgVirtual-method}
\title{Checks for existance of gene location information}
\usage{
hasGeneInfo(object)

\S4method{hasGeneInfo}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A boolean indicating whether gene location information is present 
(TRUE) or not (FALSE)
}
\description{
This method checks whether gene location information is present in the 
object i.e. if the object inherits from pgVirtualLoc
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Checks whether gene location information is available
for pgVirtual subclasses
}}

\examples{
# Exclusive pgVirtual subclasses
testPG <- .loadPgExample()
hasGeneInfo(testPG)

# pgVirtualLoc subclasses
testPG <- .loadPgExample(geneLoc=TRUE)
hasGeneInfo(testPG)

}
