% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{hasParalogueLinks}
\alias{hasParalogueLinks}
\alias{hasParalogueLinks,pgVirtual-method}
\title{Checks whether linking of paralogues has been done}
\usage{
hasParalogueLinks(object)

\S4method{hasParalogueLinks}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A boolean indicating whether paralogue links have been defined (TRUE)
or not (FALSE)
}
\description{
This method checks for the existance of paralogue links in the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Check for secondary gene grouping in pgVirtual
subclasses
}}

\examples{
# No paralogues
testPG <- .loadPgExample(withGroups=TRUE)
hasParalogueLinks(testPG)

# With paralogues
testPG <- .loadPgExample(withGroups=TRUE, withParalogues=TRUE)
hasParalogueLinks(testPG)

}
