% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R, R/pgVirtual.R
\docType{methods}
\name{internal-groupGenes}
\alias{groupGenes}
\alias{groupGenes,pgInMem-method}
\alias{groupGenes,pgVirtual-method}
\alias{internal-groupGenes}
\title{Add gene grouping to pangenome}
\usage{
groupGenes(object, ...)

\S4method{groupGenes}{pgInMem}(object, seqToGeneGroup)

\S4method{groupGenes}{pgVirtual}(object, seqToGeneGroup)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on}

\item{seqToGeneGroup}{Integer vector equivalent to the output of
seqToGeneGroup on the new object.}
}
\value{
An object with the same class as object
}
\description{
This is an internal function, not meant to be called directly. For adding
gene grouping manually see \code{\link{manualGrouping}}. This method is a
requirement for classes inheriting from pgVirtual and is not relevant for
everyday users.
}
\note{
For internal use only. Required for extending the class system of
FindMyFriends

Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}

