% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R, R/pgLM.R, R/pgLMLoc.R, R/pgFull.R, R/pgFullLoc.R
\docType{methods}
\name{internal-mergePangenomes}
\alias{internal-mergePangenomes}
\alias{mergePangenomes}
\alias{mergePangenomes,pgFull,pgFull-method}
\alias{mergePangenomes,pgFullLoc,pgFullLoc-method}
\alias{mergePangenomes,pgInMem,pgInMem-method}
\alias{mergePangenomes,pgLM,pgLM-method}
\alias{mergePangenomes,pgLMLoc,pgLMLoc-method}
\title{Merge information from two pangenomes}
\usage{
mergePangenomes(pg1, pg2, ...)

\S4method{mergePangenomes}{pgInMem,pgInMem}(pg1, pg2, geneGrouping, groupInfo,
  ...)

\S4method{mergePangenomes}{pgLM,pgLM}(pg1, pg2, geneGrouping, groupInfo, ...)

\S4method{mergePangenomes}{pgLMLoc,pgLMLoc}(pg1, pg2, geneGrouping, groupInfo)

\S4method{mergePangenomes}{pgFull,pgFull}(pg1, pg2, geneGrouping, groupInfo,
  ...)

\S4method{mergePangenomes}{pgFullLoc,pgFullLoc}(pg1, pg2, geneGrouping,
  groupInfo)
}
\arguments{
\item{pg1}{A pgVirtual subclass}

\item{pg2}{An object of the same class as pg1}

\item{...}{Parameters to be passed on}

\item{geneGrouping}{The grouping of the genees in the merged pangenome.
Equivalent to calling seqToGeneGroup on the new object}

\item{groupInfo}{The metadata on the gene groups in the merged pangenome. 
Equivalent to calling groupInfo on the new object}
}
\value{
An object of the same class as pg1
}
\description{
This method is for internal use only and should not be called directly. Use
\code{\link{addGenomes}} instead. It is required for subclasses of pgVirtual.
}
\note{
For internal use only. Required for extending the class system of 
FindMyFriends.
}
\keyword{internal}

