% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{internal-metadata}
\alias{internal-metadata}
\alias{setGroupInfo}
\alias{setGroupInfo,pgInMem-method}
\alias{setOrgInfo}
\alias{setOrgInfo,pgInMem-method}
\title{Add metadata to the pangenome}
\usage{
setGroupInfo(object, ...)

setOrgInfo(object, ...)

\S4method{setGroupInfo}{pgInMem}(object, name, info, key)

\S4method{setOrgInfo}{pgInMem}(object, name, info, key)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on}

\item{name}{The name of the metadata to set}

\item{info}{A vector of metadata}

\item{key}{The indexes the metadata in info pertains to}
}
\value{
An object of the same class as object
}
\description{
These methods are only for internal use and not relevant for regular users.
They are required for subclasses of pgVirtual and allows FindMyFriends to add
and change metadata as part of the pipelines.
}
\note{
For internal use only. Use \code{\link{addGroupInfo}} and
\code{\link{addOrgInfo}} instead. Required for extending the class system of
FindMyFriends.
}
\keyword{internal}

