% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/similarities.R
\docType{methods}
\name{kmerSimilarity}
\alias{kmerSimilarity}
\alias{kmerSimilarity,pgVirtual-method}
\title{Calculate a similarity matrix based on kmers}
\usage{
kmerSimilarity(object, ...)

\S4method{kmerSimilarity}{pgVirtual}(object, lowMem, kmerSize, lowerLimit,
  rescale, transform, pParam, nSplits)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on.}

\item{lowMem}{logical. Should low memory footprint be ensured over
computation speed}

\item{kmerSize}{The size of kmers to use for similarity calculations.}

\item{lowerLimit}{The lower threshold for similarity below which it is set to
0}

\item{rescale}{Should Similarities be normalised between lowerLimit and 1}

\item{transform}{Transformation function to apply to similarities}

\item{pParam}{An optional BiocParallelParam object that defines the workers
used for parallelisation.}

\item{nSplits}{The number of jobs to split the computations into}
}
\value{
A matrix (sparse or normal) with cosine similarity for each gene pair
}
\description{
This method takes a pangenome and calculate a similarity matrix based on
cosine similarity of kmer feature vectors in an all-vs-all fashion. The
result can subsequently be used to group genes either using
\code{\link{graphGrouping}} or homebrewed grouping scheme. In case of the
latter \code{\link{manualGrouping}} should be used to add the grouping back
to the pangenome.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Kmer based similarities for pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample()

kmerSim <- kmerSimilarity(testPG, lowerLimit=0.75)
}

