% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/splitting.R
\docType{methods}
\name{kmerSplit}
\alias{kmerSplit}
\alias{kmerSplit,pgVirtual-method}
\title{Split gene groups based on similarity}
\usage{
kmerSplit(object, ...)

\S4method{kmerSplit}{pgVirtual}(object, kmerSize, lowerLimit, maxLengthDif,
  pParam)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Arguments passed on}

\item{kmerSize}{The length of kmers used for sequence similarity}

\item{lowerLimit}{The lower limit of sequence similarity below which it will
be set to 0}

\item{maxLengthDif}{The maximum deviation in sequence length to allow. 
Between 0 and 1 it describes a percentage. Above 1 it describes a fixed 
length}

\item{pParam}{An optional BiocParallelParam object that defines the workers 
used for parallelisation.}
}
\value{
A new pgVirtual subclass object of the same class as 'object'
}
\description{
This function splits up gene groups based on cosine similarity of kmer 
feature vectors. It uses hard splitting based on a similarity cutoff where
unconnected components constitutes new groups. Unlike 
\code{\link{neighborhoodSplit}}, paralogues cannot be forced into separate 
groups as information needed for this is not present.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Kmer similarity based group splitting for pgVirtual 
subclasses
}}
\examples{
# Get a grouped pangenome
pg <- .loadPgExample(withGroups = TRUE)

# Split groups by similarity
pg <- kmerSplit(pg, lowerLimit = 0.8)

}
\seealso{
Other group-splitting: \code{\link{neighborhoodSplit}}
}

