% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.loadPgExample}
\alias{.loadPgExample}
\title{Load an example pangenome}
\usage{
.loadPgExample(lowMem = FALSE, geneLoc = FALSE, withGroups = FALSE,
  withNeighborhoodSplit = FALSE, withParalogues = FALSE)
}
\arguments{
\item{lowMem}{logical. Should the returned object inherit from pgLM}

\item{geneLoc}{logical. Should the returned object inherit from pgVirtualLoc}

\item{withGroups}{logical. Should gene groups be defined}

\item{withNeighborhoodSplit}{logical. Should neighborhoodsplitting have been
performed}

\item{withParalogues}{logical. Should paralogue linking have been performed}
}
\value{
A pgVirtual subclass object to the specifications defined
}
\description{
This function loads an example pangenome at various stages of calculation,
useful for examples and tests.
}
\examples{
# Load standard (pgFull)
.loadPgExample()

# Use pgLM
.loadPgExample(lowMem=TRUE)

# Create with pgVirtualLoc subclass (here pgFullLoc)
.loadPgExample(geneLoc=TRUE)

# Create with grouping information
.loadPgExample(withGroups=TRUE)

# Create with gene groups split by neighborhood (pgVirtualLoc implied)
.loadPgExample(withNeighborhoodSplit=TRUE)

# Create with paralogue links
.loadPgExample(withGroups=TRUE, withParalogues=TRUE)

}

