% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grouping.R
\docType{methods}
\name{manualGrouping}
\alias{manualGrouping}
\alias{manualGrouping,pgVirtual,integer-method}
\alias{manualGrouping,pgVirtual,list-method}
\title{Define gene grouping manually}
\usage{
manualGrouping(object, groups)

\S4method{manualGrouping}{pgVirtual,integer}(object, groups)

\S4method{manualGrouping}{pgVirtual,list}(object, groups)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{groups}{Either a list or integer vector defining the grouping}
}
\value{
An object of the same class as 'object'.
}
\description{
In cases where results from other algorithms are wished to be imported into
the FindMyFriends framework, this method ensures that the proper formatting 
is done. The grouping can  be defined as an integer vector with an element 
for each gene. The value of each element is then used as the gene group 
classifier. Alternatively groups can be defined by a list of integer vectors.
Each element of the list defines a group and the content of each element 
refers to gene indexes.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = pgVirtual,groups = integer}: manual grouping defined by integer vector

\item \code{object = pgVirtual,groups = list}: manual grouping defined by list
}}

\examples{
testPG <- .loadPgExample()

# Load grouping data
groups <- system.file('extdata', 'examplePG', 'groupsWG.txt', 
    package='FindMyFriends'
)
groups <- scan(groups, what=integer(), quiet=TRUE)

# Do the grouping
testPG <- manualGrouping(testPG, groups)

}
\seealso{
Other grouping algorithms: \code{\link{cdhitGrouping}},
  \code{\link{gpcGrouping}}, \code{\link{graphGrouping}}
}
\concept{grouping algorithms}
