% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/splitting.R
\docType{methods}
\name{neighborhoodSplit}
\alias{neighborhoodSplit}
\alias{neighborhoodSplit,pgVirtualLoc-method}
\title{Split gene groups by neighborhood synteny}
\usage{
neighborhoodSplit(object, ...)

\S4method{neighborhoodSplit}{pgVirtualLoc}(object, flankSize, minFlank,
  forceParalogues, kmerSize, lowerLimit, maxLengthDif)
}
\arguments{
\item{object}{A pgVirtualLoc subclass}

\item{...}{parameters passed on.}

\item{flankSize}{The number of flanking genes on each side of the gene to use
for comparison.}

\item{minFlank}{The lowest number of matching genes on either side to allow}

\item{forceParalogues}{Force similarity of paralogue genes to 0}

\item{kmerSize}{The length of kmers used for sequence similarity}

\item{lowerLimit}{The lower limit of sequence similarity below which it will
be set to 0}

\item{maxLengthDif}{The maximum deviation in sequence length to allow.
Between 0 and 1 it describes a percentage. Above 1 it describes a fixed
length}
}
\value{
An object with the same class as object containing the new grouping.
}
\description{
This function evaluates already created gene groups and splits the members
into new groups based on the synteny of the flanking genes and the similarity
of the sequences. In general the splitting is based on multiple stages that
all gene pairs must pass in order to remain in the same group. First the link
between the genes is removed if they are part of the same organism. Then the
synteny of the flanking genes are assessed and if it doesn't passes the
defined threshold the link between the gene pair is removed. Then the kmer
similarity of the two sequences are compared and if below a certain threshold
the link is removed. Lastly the length of the two sequences are compared and
if below a certain threshold the link is removed. Based on this new graph
cliques are detected and sorted based on the lowest within-clique sequence
similarity and neighborhood synteny. The cliques are then added as new groups
if the members are not already members of a new group until all members are
part of a new group. This approach ensures that all members of the new
groupings passes certain conditions when compared to all other members of the
same group.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtualLoc}: Neighborhood-based gene group splitting for
pgVirtualLoc subclasses
}}
\examples{
testPG <- .loadPgExample(geneLoc=TRUE, withGroups=TRUE)

testPG <- neighborhoodSplit(testPG, lowerLimit=0.75)
}

