% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{nGeneGroups}
\alias{nGeneGroups}
\alias{nGeneGroups,pgVirtual-method}
\title{Get the number of gene groups in a pangenome}
\usage{
nGeneGroups(object)

\S4method{nGeneGroups}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
An integer giving the number of gene groups
}
\description{
This method gives the number of different gene groups in the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: The number of gene groups in the pangenome for
pgVirtual subclasses
}}

\examples{
testPG <- .loadPgExample(withGroups=TRUE)
nGeneGroups(testPG)

}
