% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{nGenes}
\alias{nGenes}
\alias{nGenes,pgVirtual-method}
\title{Get the total number of genes in a pangenome}
\usage{
nGenes(object)

\S4method{nGenes}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
An integer giving the number of genes in the object
}
\description{
This method returns the total number of genes in a pangenome (i.e. the sum
of genes in each organism in the pangenome)
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: The number of genes in the pangenome for pgVirtual 
subclasses.
}}

\examples{
testPG <- .loadPgExample()
nGenes(testPG)

}
