% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{orgInfo}
\alias{orgInfo}
\alias{orgInfo,pgInMem-method}
\alias{orgInfo<-}
\alias{orgInfo<-,pgInMem-method}
\title{Get and set information about organisms}
\usage{
orgInfo(object)

orgInfo(object) <- value

\S4method{orgInfo}{pgInMem}(object)

\S4method{orgInfo}{pgInMem}(object) <- value
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{value}{A data.frame with a row for each organism}
}
\value{
In case of the getter a data.frame with organism information.
}
\description{
These methods lets you access the information stored about each organism and
add to it or modify it. The only information present up front is the number 
of genes present in each organism. While possible, this information should 
not be changed manually but through the \code{\link{removeGene}} functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Get organism metadata for pgInMem subclasses

\item \code{pgInMem}: Set organism metadata for pgInMem subclasses
}}
\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample()
orgInfo(testPG)

orgInfo(testPG)$Genus <- 'Mycoplasma'

}
\seealso{
Other Metadata: \code{\link{addGroupInfo}},
  \code{\link{addOrgInfo}}, \code{\link{groupInfo}}
}

