% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{orgNames}
\alias{orgNames}
\alias{orgNames<-}
\alias{orgNames,pgInMem-method}
\alias{orgNames<-,pgInMem-method}
\title{Get and set the names of organisms in the pangenome}
\usage{
orgNames(object)

orgNames(object) <- value

\S4method{orgNames}{pgInMem}(object)

\S4method{orgNames}{pgInMem}(object) <- value
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{value}{A vector with new names - will be coerced to characters}
}
\value{
In case of the getter a character vector with names
}
\description{
These methods lets you manipulate the naming of organisms in the pangenome.
By default organisms are named after the fasta file they are defined by, but
this can be changed at will.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Get organism names for pgInMem subclasses

\item \code{pgInMem}: Set organism names for pgInMem subclasses
}}

\note{
Required for subclasses of pgVirtual in order to extend the class
system of FindMyFriends
}
\examples{
testPG <- .loadPgExample()
orgNames(testPG)

orgNames(testPG)[3] <- 'Organism 3'

}
