% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/investigating.R
\docType{methods}
\name{orgStat}
\alias{orgStat}
\alias{orgStat,pgVirtual-method}
\title{Calculate statistics about each organism}
\usage{
orgStat(object, ...)

\S4method{orgStat}{pgVirtual}(object, subset, getFrequency = FALSE)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{parameters passed on.}

\item{subset}{Name or indexes of organisms to include}

\item{getFrequency}{logical. Should amino/nucleic acid frequency be
calculated}
}
\value{
A data.frame with a row per organism, with each statistic in a column
}
\description{
This method, much like {code{\link{groupStat}}} calculates different
statistics for each organism in the pangenome. Depending on the parameters
the statistics are: number of genes, minimum length of gene, maximum length
of gene standard deviation of gene lengths, residue frequency, number of gene
groups and number of paralogues.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Organism statistics for all pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

orgStats <- orgStat(testPG)
}

