% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructor.R
\name{pangenome}
\alias{pangenome}
\title{Construct a pangenome from fasta files}
\usage{
pangenome(paths, translated, geneLocation = NULL, lowMem = FALSE, ...)
}
\arguments{
\item{paths}{A character vector with location of fasta files}

\item{translated}{A boolean indicating if the fasta files contain amino acid
sequences}

\item{geneLocation}{A function, string or dataframe. If it is a data.frame it
should contain the columns 'contig', 'start', 'end' and 'strand' with a row
for each gene. If it is a function it should take the name (fasta
description) for each gene and output a data.frame similar to described
above. If it is a string it should specify the format of the gene names.
Currently only 'prodigal' is supported.}

\item{lowMem}{Boolean. Should FindMyFriends avoid storing sequences in
memory.}

\item{...}{Additional defaults to set on the object}
}
\value{
A pgVirtual subclass object depending on geneLocation and lowMem.
\tabular{lll}{
 \bold{geneLocation} \tab \bold{lowMem} \tab \bold{Resulting class}         \cr
 NULL                \tab FALSE         \tab \code{\linkS4class{pgFull}}    \cr
 NULL                \tab TRUE          \tab \code{\linkS4class{pgLM}}      \cr
 !NULL               \tab FALSE         \tab \code{\linkS4class{pgFullLoc}} \cr
 !NULL               \tab TRUE          \tab \code{\linkS4class{pgLMLoc}}   \cr
}
}
\description{
This function constructs an initial pangenome object from a set of fasta
files. Note that the actual pangenome is not calculated here. As such this
function mainly sets everything up before beginning the more lengthly
pangenome calculation.
}
\examples{
location <- tempdir()
unzip(system.file('extdata', 'Mycoplasma.zip', package='FindMyFriends'),
      exdir=location)
genomeFiles <- list.files(location, full.names=TRUE, pattern='*.fasta')

# Create pgFull
pangenome(genomeFiles, TRUE)

# Create pgFullLoc
pangenome(genomeFiles, TRUE, geneLocation='prodigal')

# Create pgLM
pangenome(genomeFiles, TRUE, lowMem=TRUE)

# Create pgLMLoc
pangenome(genomeFiles, TRUE, geneLocation='prodigal', lowMem=TRUE)
}

