% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/investigating.R
\docType{methods}
\name{pcGraph}
\alias{pcGraph}
\alias{pcGraph,pgVirtualLoc-method}
\title{Calculate the panchromosome graph}
\usage{
pcGraph(object, ...)

\S4method{pcGraph}{pgVirtualLoc}(object, slim = FALSE)
}
\arguments{
\item{object}{A pgVirtualLoc subclass}

\item{slim}{Should the returned graph be stripped of all metadata and only
capture gene group connectivity. Defaults to FALSE}
}
\value{
An igraph object
}
\description{
This method creates a graph representation of the panchromosome - The 
complete set of gene groups linked together by chromosomal position. Each 
vertice in the graph represent a gene group and each edge represent a 
positional relation between two gene groups (neighboring each other). 
Vertices are annotated with number of genes, organism names and strand while
edges are annotated with numer of genes (as weight), and organism names.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtualLoc}: Panchromosome creation for all pgVirtualLoc subclasses
}}
\examples{
testPG <- .loadPgExample(geneLoc=TRUE, withNeighborhoodSplit=TRUE)

panchromosome <- pcGraph(testPG)

}

