% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgFull.R
\docType{class}
\name{pgFull-class}
\alias{pgFull-class}
\title{Class for in memory pangenome data}
\description{
This class handles pangenome data without gene location information and with
all sequences stored in memory. This makes sequence lookup much faster but
also increases the memory footprint of the object thus making it a bad choice
for very large pangenome with millions of genes.
}
\section{Slots}{

\describe{
\item{\code{sequences}}{Either an AAStringSet or DNAStringSet containing all 
sequences in the pangenome.}
}}
\seealso{
Other Pangenome_classes: \code{\link{pgFullLoc-class}},
  \code{\link{pgInMem-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}}, \code{\link{pgSlim-class}},
  \code{\link{pgSlimLoc-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}

