% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgFullLoc.R
\docType{class}
\name{pgFullLoc-class}
\alias{pgFullLoc-class}
\title{Class for in memory pangenome data with location information}
\description{
This class extends \code{\linkS4class{pgFull}} by subclassing 
\code{\linkS4class{pgInMemLoc}} and thus adding gene location information
to each gene. See the respective superclasses for more information.
}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgInMem-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}}, \code{\link{pgSlim-class}},
  \code{\link{pgSlimLoc-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}

