% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInMem.R
\docType{class}
\name{pgInMem-class}
\alias{pgInMem-class}
\title{FindMyFriends standard base class for pangenomic data}
\description{
This virtual class is the superclass of the standard pangenome classes in
FindMyFriends. It defines storage for everything except gene information,
which is delegated to its subclasses.
}
\details{
As gene storage is not defined in this class the following methods must be
defined by subclasses:

\describe{
 \item{genes(object, split, subset)}{Return the underlying sequences. If 
 split is missing return an XStringSet, otherwise return an XStringSetList. 
 split can be either 'group', 'organism' or 'paralogue' and should group the 
 sequences accordingly. Subset should behave as if it was added as '[]' to 
 the results but allow you to avoid reading everything into memory if not 
 needed.}
 \item{geneNames(object)}{Return a character vector with the name of each 
 gene.}
 \item{geneNames<-(object, value)}{Set the name of each gene.}
 \item{geneWidth(object)}{Return an integer vector with the length (in 
 residues) of each gene.}
 \item{removeGene(object, name, organism, group, ind)}{\strong{Should only be
 implemented for signature: c(\emph{yourClass}, 'missing', 'missing', 
 'missing', 'integer')} Remove the genes at the given indexes and return the 
 object.}
}
}
\section{Slots}{

\describe{
\item{\code{seqToOrg}}{An integer vector that reference all genes to a specific 
organism.}

\item{\code{seqToGeneGroup}}{An integer vector that references all genes to a 
specific gene group.}

\item{\code{groupInfo}}{A data.frame storing metadata information about gene groups.}

\item{\code{orgInfo}}{A data.frame storing metadata information about organisms}
}}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgFullLoc-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}

