% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInMemLoc.R
\docType{class}
\name{pgInMemLoc-class}
\alias{pgInMemLoc-class}
\title{Superclass for gene location aware pangenome}
\description{
This virtual class is the superclass for all standard, location aware, 
pangenome classes in FindMyFriends. It stores all chromosomal information in
a data.frame.
}
\section{Slots}{

\describe{
\item{\code{geneLocation}}{A data.frame containing the columns 'contig', 'start', 
'end' and 'strand' and a row for each gene in the pangenome.}
}}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgFullLoc-class}},
  \code{\link{pgInMem-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}}, \code{\link{pgSlim-class}},
  \code{\link{pgSlimLoc-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}

