% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgSlim.R
\docType{class}
\name{pgSlim-class}
\alias{pgSlim-class}
\title{Class for pangenome data with no reference to genes}
\description{
This class is a slim version of pgLM and pgFull that does not store any
information pertaining to the actual genes. This means that this class cannot
be the basis for the creation of a pangenome but that pgLM or pgFull objects
can be coerced down to this representation after the pangenome has been 
created to make it less burdensome to work with, while still keeping a lot
of the functionality of the FindMyFriends framework.
}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgFullLoc-class}},
  \code{\link{pgInMem-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}},
  \code{\link{pgSlimLoc-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}
\concept{Pangenome_classes}
