% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgSlimLoc.R
\docType{class}
\name{pgSlimLoc-class}
\alias{pgSlimLoc-class}
\title{Class for pangenome data with no reference to genes}
\description{
This class extends \code{\linkS4class{pgSlim}} by subclassing 
\code{\linkS4class{pgInMemLoc}} and thus adding gene location information
to each gene. See the respective superclasses for more information.
}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgFullLoc-class}},
  \code{\link{pgInMem-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}}, \code{\link{pgSlim-class}},
  \code{\link{pgVirtual-class}},
  \code{\link{pgVirtualLoc-class}}
}

