% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgVirtualLoc.R
\docType{class}
\name{pgVirtualLoc-class}
\alias{pgVirtualLoc-class}
\title{Superclass for gene location aware pangenome}
\description{
This virtual class should be subclassed by all classes that include 
chromosomal position of the genes (along with subclassing pgVirtual). The
class itself is an empty shell that only takes care of dispatching and 
checking the promises of subclasses are held.
}
\details{
Subclasses of pgVirtualLoc must implement the following methods:

\describe{
 \item{geneLocation(object)}{Return a data.frame with a row for each gene,
 describing the chromosomal position of the gene. The data.frame must contain
 the columns 'contig', 'start', 'end' and 'strand'. Contig is 
 self-explanatory, start and end is the respective start and end positions on
 the contig (start must be lower than end) and strand defines the coding 
 direction as -1 or 1.}
}
}
\seealso{
Other Pangenome_classes: \code{\link{pgFull-class}},
  \code{\link{pgFullLoc-class}},
  \code{\link{pgInMem-class}},
  \code{\link{pgInMemLoc-class}}, \code{\link{pgLM-class}},
  \code{\link{pgLMLoc-class}},
  \code{\link{pgVirtual-class}}
}

