% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{plotEvolution}
\alias{plotEvolution}
\alias{plotEvolution,pgVirtual-method}
\title{Plot the evolution in gene groups}
\usage{
plotEvolution(object, ...)

\S4method{plotEvolution}{pgVirtual}(object, ordering = "bootstrap",
  times = 10)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on}

\item{ordering}{An ordering of the organisms by name or index, or alternative
one of 'bootstrap', 'random' or 'none'.}

\item{times}{The number of sampling for ordering='bootstrap'}
}
\description{
This method constructs a plot showing how the number of singleton, accessory
and core gene groups evolve as the size of the pangenome increases. Different
ways of increasing the size of the pangenome is available.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Evolution plots for pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Standard type - organisms ordered by their index in the pangenome
plotEvolution(testPG, ordering='none')

# Bootstrapped with confidence intervals
plotEvolution(testPG, ordering='bootstrap')
}

