% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/investigating.R
\docType{methods}
\name{plotGroup}
\alias{plotGroup}
\alias{plotGroup,pgVirtual-method}
\title{Plot the similarities of genes within a group}
\usage{
plotGroup(object, ...)

\S4method{plotGroup}{pgVirtual}(object, group, kmerSize, lowerLimit, rescale,
  transform, ...)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on to igraphs plotting method}

\item{group}{Name or index of the gene group to plot}

\item{kmerSize}{The kmer size to use for similarity calculations}

\item{lowerLimit}{The lower threshold for similarity below which it will be
set to 0}

\item{rescale}{logical. Should the similarity be rescaled between lowerLimit
and 1}

\item{transform}{A transformation function to apply to the similarities}
}
\value{
Called for the side effect of creating a plot. Invisibly returns an 
igraph object with all visual parameters set as node and edge attributes.
}
\description{
This method plots a gene group with genes as vertices and cosine similarities
as weighted edges. Mildly informative at best :-)
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Gene group similiarity plotting for all pgVirtual
subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

plotGroup(testPG, 10, lowerLimit=0.25)

}

