% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/investigating.R
\docType{methods}
\name{plotNeighborhood}
\alias{plotNeighborhood}
\alias{plotNeighborhood,pgVirtualLoc-method}
\title{Plot the neighborhood of a gene group}
\usage{
plotNeighborhood(object, ...)

\S4method{plotNeighborhood}{pgVirtualLoc}(object, group, vicinity = 4, ...)
}
\arguments{
\item{object}{A pgVirtualLoc subclass}

\item{...}{Parameter passed on to igraph's plot method.}

\item{group}{The name or index of a group.}

\item{vicinity}{An integer giving the number of gene groups in both 
directions to collect.}
}
\value{
Called for the side effect of creating a plot. Invisibly returns an 
igraph object with all visual parameters set as node and edge attributes.
}
\description{
This method plots the neighborhood extracted using 
\code{\link{getNeighborhood}} in a visually pleasing way. It is mainly a 
wrapper around \code{\link[igraph]{plot.igraph}} to ensure the proper 
information is visualised.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtualLoc}: Gene group neighborhood plotting for all
pgVirtualLoc subclasses
}}
\examples{
testPG <- .loadPgExample(geneLoc=TRUE, withNeighborhoodSplit=TRUE)

# Nice little overview of the neighborhood of gene group 30
plotNeighborhood(testPG, 30)

}

