% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{plotSimilarity}
\alias{plotSimilarity}
\alias{plotSimilarity,pgVirtual-method}
\title{Create a heatplot with similarities between all organisms}
\usage{
plotSimilarity(object, ...)

\S4method{plotSimilarity}{pgVirtual}(object, type = "pangenome",
  ordering = "auto", kmerSize, pParam, chunkSize = 100)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on.}

\item{type}{The type of similarity calculation. Either 'pangenome' or 'kmer'}

\item{ordering}{The ordering of rows and column in the heatmap. Either 
integer og character vector with organism names or one of the following:
'auto' or 'none'. For 'auto' The ordering will be based on a hierarchical 
clustering of the organisms based on Ward's distance.}

\item{kmerSize}{The size of the kmers to use for comparison}

\item{pParam}{An object of class BiocParallelParam}

\item{chunkSize}{Number of organisms to process at a time}
}
\value{
This function is called for its side effects
}
\description{
This method creates a heatplot showing the similarity between all organisms
in the pangenome. The similarity can either be derived from the pangenome
matrix or from kmer calculations of the genes themselves.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Similarity heatmaps for pgVirtual subclasses
}}

\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Use kmers
plotSimilarity(testPG, type='kmer')

# Use pangenome matrix
plotSimilarity(testPG, type='pangenome')

}
\seealso{
\code{\link{plotTree}} for a dendrogram plot of the same data.
}
