% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{plotStat}
\alias{plotStat}
\alias{plotStat,pgVirtual-method}
\title{Plot (very) basic statistics on the pangenome}
\usage{
plotStat(object, ...)

\S4method{plotStat}{pgVirtual}(object, sort = TRUE, color, ...)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters passed on to color scale.}

\item{sort}{logical. Should Genomes be sorted based on their number of genes}

\item{color}{A metadata name to color the organisms by}
}
\value{
This function is called for its side effects
}
\description{
This method plots the number of genes in each organism and, if gene groups
have been defined, the number of singleton, accessory and core gene groups.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Plot basic statistics for pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Should make a nice little plot
plotStat(testPG)
}

