% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{plotTree}
\alias{plotTree}
\alias{plotTree,pgVirtual-method}
\title{Plot a dendrogram of the organisms in a pangenome}
\usage{
plotTree(object, ...)

\S4method{plotTree}{pgVirtual}(object, type = "pangenome", circular = FALSE,
  info, kmerSize, dist, clust, pParam, chunkSize = 100)
}
\arguments{
\item{object}{A pgVirtual subclass}

\item{...}{Parameters to be passed on.}

\item{type}{The type of distance meassure. Either 'pangenome' or 'kmer'}

\item{circular}{logical. Should the dendrogram be plotted in a circular
fashion.}

\item{info}{Metadata to plot at the leafs of the dendrogram. Either the name
of an orgInfo column or a vector with information for each organism.}

\item{kmerSize}{The size of the kmers to use for comparison}

\item{dist}{The distance function to use. All possible values of method in 
dist() are allowed as well as 'cosine' for type='kmer'}

\item{clust}{The clustering function to use. Passed on to method in hclust()}

\item{pParam}{An object of class BiocParallelParam}

\item{chunkSize}{Number of organisms to process at a time}
}
\value{
This function is called for its side effects
}
\description{
This method plots a dendrogram of the relationship between the organisms in
the pangenome. It does not tries to by phylogenetic in any way but merely 
shows the relationship in data. As with \code{\link{plotSimilarity}} it can
be based on either the pangenome matrix or kmer feature vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Dendrogram plotting for pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

plotTree(testPG, type='pangenome', dist='binary', clust='ward.D2')

# And now in a circle (type defaults to 'pangenome')
plotTree(testPG, circular=TRUE, dist='binary', clust='ward.D2')

}
\seealso{
\code{\link{plotSimilarity}} for a heatmap plot of the same data.
}

