% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa.R
\name{readAnnot}
\alias{readAnnot}
\title{Import annotation from an .annot file}
\usage{
readAnnot(file)
}
\arguments{
\item{file}{The .annot file to import}
}
\value{
A data.frame ready to merge with a pangenome object using
\code{\link{addGroupInfo}} with the \code{key} argument set to 'name'.
}
\description{
This function makes it easy to import annotation create in Blast2GO or other
programs supporting .annot exporting of results.
}
\examples{
# Get path to file
annot <- system.file('extdata', 'examplePG', 'example.annot', package='FindMyFriends')

# Parse the file
readAnnot(annot)
}

