% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifying.R
\name{reportGroupChanges}
\alias{reportGroupChanges}
\title{Reports the change in grouping}
\usage{
reportGroupChanges(newGrouping, oldGrouping, file)
}
\arguments{
\item{newGrouping}{An integer vector as produced by 
\code{\link{seqToGeneGroup}} with the grouping after the change}

\item{oldGrouping}{An integer vector as produced by 
\code{\link{seqToGeneGroup}} with the grouping before the change}

\item{file}{A file to write}
}
\value{
This function is called for its side effects
}
\description{
This function inspects gene grouping before and after a change and reports on
the changes. If newGrouping is missing it reports on the last performed 
comparison; optionally writing it to a file if 'file' is specified.
}
\examples{
# Show latest changes in grouping
reportGroupChanges()

# Alternatively write it to a file
reportGroupChanges(file = tempfile())

}
