% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{seqToGeneGroup}
\alias{seqToGeneGroup}
\alias{seqToGeneGroup,pgInMem-method}
\title{Get gene-to-genegroup relationship}
\usage{
seqToGeneGroup(object)

\S4method{seqToGeneGroup}{pgInMem}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
An integer vector with an element for each gene in the pangenome.
}
\description{
This method returns the group membership for each gene in the pangenome as
a vector of indices. Element 1 corresponds to gene 1 and the value is the
index of the corresponding gene group. If gene groups have yet to be defined
it returns a vector of length 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Gene to genegroup indexing for pgInMem subclasses
}}

\note{
Required for extending the class system of FindMyFriends
}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Have a look at what the first six genes belongs to
head(seqToGeneGroup(testPG))

}
\seealso{
\code{\link{seqToOrg}} for gene-to-organism relationship
}
