% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgInMem.R
\docType{methods}
\name{seqToOrg}
\alias{seqToOrg}
\alias{seqToOrg,pgInMem-method}
\title{Get gene-to-organism relationship}
\usage{
seqToOrg(object)

\S4method{seqToOrg}{pgInMem}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
An integer vector with an element for each gene in the pangenome.
}
\description{
This method returns the organism membership for each gene in the pangenome as
a vector of indices. Element 1 corresponds to gene 1 and the value is the
index of the corresponding organism.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgInMem}: Gene to organism indexing for pgInMem subclasses
}}

\note{
Required for extending the class system of FindMyFriends
}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

# Stored sequentially so the first will belong to organism 1
head(seqToOrg(testPG))

}
\seealso{
\code{\link{seqToGeneGroup}} for gene-to-genegroup relationship
}
