% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{translated}
\alias{translated}
\alias{translated,pgVirtual-method}
\title{Check the sequence type of the pangenome}
\usage{
translated(object)

\S4method{translated}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A boolean indicating whether genes are translated (TRUE) or not 
(FALSE)
}
\description{
This method checks whether the genes in the pangenome are on translated
form (amino acid sequences) or not. A return value of FALSE only indicates
that the storage mode for the genes is not an AAStringSet. While this leaves
room for both RNA-, DNA- and BStringSet, only DNAStringSet makes much sense
and is therefore assumed
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Sequence type check for pgVirtual subclasses
}}

\examples{
testPG <- .loadPgExample()

# Genes are translated
translated(testPG)

# ... and therefore returned as AAStringSet instead of DNAStringSet
class(genes(testPG, subset=1))

}
