% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pgInMemLoc.R
\docType{class}
\name{pgInMemLoc-class}
\alias{pgInMemLoc-class}
\title{Superclass for gene location aware pangenome}
\description{
This virtual class is the superclass for all standard, location aware,
pangenome classes in FindMyFriends. It stores all chromosomal information in
a data.frame.
}
\section{Slots}{

\describe{
\item{\code{geneLocation}}{A data.frame containing the columns 'contig', 'start',
'end' and 'strand' and a row for each gene in the pangenome.}
}}
\seealso{
Other Pangenome_classes: \code{\link{as}},
  \code{\link{as}}, \code{\link{length,pgVirtual-method}},
  \code{\link{pgVirtual-class}},
  \code{\link{show,pgVirtual-method}};
  \code{\link{pgFull-class}};
  \code{\link{pgFullLoc-class}};
  \code{\link{pgInMem-class}}; \code{\link{pgLM-class}};
  \code{\link{pgLMLoc-class}};
  \code{\link{pgVirtualLoc-class}}
}

