% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pgLMLoc.R
\docType{class}
\name{pgLMLoc-class}
\alias{pgLMLoc-class}
\title{Class for reference based pangenome data with location information}
\description{
This class extends \code{\linkS4class{pgLM}} by subclassing
\code{\linkS4class{pgVirtualLoc}} and thus adding gene location information
to each gene. See the respective superclasses for more information.
}
\seealso{
Other Pangenome_classes: \code{\link{as}},
  \code{\link{as}}, \code{\link{length,pgVirtual-method}},
  \code{\link{pgVirtual-class}},
  \code{\link{show,pgVirtual-method}};
  \code{\link{pgFull-class}};
  \code{\link{pgFullLoc-class}};
  \code{\link{pgInMem-class}};
  \code{\link{pgInMemLoc-class}}; \code{\link{pgLM-class}};
  \code{\link{pgVirtualLoc-class}}
}

