% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pgVirtual.R
\docType{methods}
\name{pgMatrix}
\alias{pgMatrix}
\alias{pgMatrix,pgVirtual-method}
\title{Get the pangenome matrix}
\usage{
pgMatrix(object)

\S4method{pgMatrix}{pgVirtual}(object)
}
\arguments{
\item{object}{A pgVirtual subclass}
}
\value{
A matrix with organisms as columns and gene groups as rows
}
\description{
This method lets you extract the pangenome matrix of the pangenome. It is not
possible to directly change the pangenome matrix as it not necessary stored 
in the object but might be calculated on request. Either way the pangenome
matrix is a function of the gene grouping and should be changed by changing 
the gene grouping instead of being manipulated downstream.
}
\section{Methods (by class)}{
\itemize{
\item \code{pgVirtual}: Get pangenome matrix for pgVirtual subclasses
}}
\examples{
testPG <- .loadPgExample(withGroups=TRUE)

head(pgMatrix(testPG))

}

